pyproject.toml
[project]
name = "coverlettercreator"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "Sam Siggs", email = "sam.siggs@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain==0.3.7",
    "langchain-community==0.3.7",
    "langchain-core (>=0.3.34,<0.4.0)",
    "tiktoken==0.8.0",
    "chromadb==0.5.20",
    "onnxruntime==1.20.0",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.61.1,<2.0.0)",
    "langchain-openai (>=0.3.4,<0.4.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "reportlab (>=4.3.0,<5.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "playwright (>=1.51.0,<2.0.0)",
    "playwright-stealth (>=1.0.6,<2.0.0)",
    "setuptools (>=78.1.0,<79.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

